Work In Progress  Environment: * Don't increment head_id, just pop from the front, append to the end and always access the first path in the list. Assign paths termination points only once they are accessed. * Every step computes the same step target regardless of terminating. If there are no neuron nodes or termination points within the window, target point is the current position. * The reward is multiplied by 1/(1-gamma) for terminations. * Add option of adding branches every 7 pixels during training. Tracking Reward: * New get termination nodes function that avoids recursion * Fixes to target point computation * Fix to remove_visited Plotting: * Add trace_gif.py to create function for saving gifs of traces during training. Training: * Set complexity increases to every 10000 episodes * save a gif every 1000 episodes

+++ bin/setup_sac_train_v1.py
@@ -56,7 +56,7 @@ def setup(
         use_symlinks: Whether to use symlinks for output files.
         remove_soma: Whether to remove soma from SWC.
     """
-    if remove_soma:
+    if remove_soma and swc_dir is not None:
         for swc_file in list(Path(swc_dir).rglob("*.swc")):
             swc_list = load.swc(swc_file)
             swc_list, seeds = tree.remove_soma(swc_list, max_radius=7.0)
@@ -69,9 +69,9 @@ def setup(
     complexity_config = DrawingComplexityConfig()
     rng = np.random.default_rng(rng_seed)
     data_generator = DataGenerator(cache_dir=output_dir, complexity_config=complexity_config, rng=rng)
-    data_generator.generate_data(output_dir, subtrees_per_swc=subtrees_per_swc, complexity_range=complexity_range, swc_dir=swc_dir, img_dir=image_dir, dataset_size=dataset_size)
+    data_generator.generate_data(subtrees_per_swc=subtrees_per_swc, complexity_range=complexity_range, swc_dir=swc_dir, img_dir=image_dir, dataset_size=dataset_size)
 
-    if remove_soma:
+    if remove_soma and swc_dir is not None:
         # remove temporary directory
         shutil.rmtree(temp_swc_dir)
